# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Documentación",
    "version": "1.0.0",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://api.ochio-club.com",
      "description": "Servidor de producción"
    }
  ],
  "components": {
    "securitySchemes": {
      "apiKeyAuth": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "example": {
          "error": "BadRequest",
          "message": "Solicitud inválida"
        }
      }
    },
    "examples": {
      "ProductExample": {
        "value": {
          "id": "3",
          "name": "Producto Ejemplo",
          "price": 99.99
        }
      }
    }
  },
  "security": [
    {
      "apiKeyAuth": []
    }
  ],
  "paths": {
    "/products": {
      "get": {
        "operationId": "getProducts",
        "summary": "Obtener lista de productos",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Número de página",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Cantidad de productos por página",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de productos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "price": {
                        "type": "number"
                      }
                    }
                  }
                },
                "example": [
                  {
                    "id": "1",
                    "name": "Producto 1",
                    "price": 100
                  },
                  {
                    "id": "2",
                    "name": "Producto 2",
                    "price": 200
                  }
                ]
              }
            }
          },
          "400": {
            "description": "Solicitud inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createProduct",
        "summary": "Crear un nuevo producto",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number"
                  }
                },
                "required": ["name", "price"]
              },
              "example": {
                "$ref": "#/components/examples/ProductExample"
              }
            }
          },
          "responses": {
            "201": {
              "description": "Producto creado exitosamente",
              "content": {
                "application/json": {
                  "example": {
                    "$ref": "#/components/examples/ProductExample"
                  }
                }
              }
            },
            "400": {
              "description": "Solicitud inválida",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}